Initial work on simplified instructions
This commit lays the groundwork for unifying disparately typed versions
of instructions. Previously we had instructions such as mov_reg_u8,
mov_reg_u16, etc. which meant a lot of code reuse and complexity.
Additionally, as I made the assembler more general I found that my
previous approach was unsuitable as there were too many conditions to
consider. Further, this approach greatly simplifies the assembler as can
be seen. Rather than caring about the type of immediates, we simply
convert the AST to an Immediate and call the proper instruction. I
believe that this approach will continue to work going forward.