Add beginnings of parser
I started deciding how the language will look with this commit. Things
are starting to take shape and it has become clear to me that if this
language is going to do what I want it is going to look quite different
from Scheme. I had intended to allow ({[< to be used interchangably so
long as they matched up properly. However allowing <> (which I intended
to use for type definitions like <Vec u8>) means that a name like
i32->String will not be handled properly which is unfortunate as I am
very fond of this naming style.

For now I haven't done anything so a name like i32->String does not
work. I don't think that there is an effective way to allow them to
coexist so I will likely have to decide which I prefer.