Remove some hardcoded values
Hopefully this now supports arbitrary amounts of code/data.
Will have to add a relocation feature to insert the proper addresses for
loading constant values like strings. Probably going to change Elf.data
to a HashMap<Vec<u8>, Vec<usize>> where the first vec is the constant and the second vec is the locations of addresses in the program code which need to be changed. That way we can go through and set them to what they will be when loaded.